Fix data race in WithContext with cancelled parent
If the parent context is already cancelled, or cancelled before
WithContext returns, there is a data race that can happen between the
goroutine that is spawned to detect when the parent context is cancelled
or the tomb is dying, which calls Kill() and the code in WithContext
that updates the children.

This change fixes the race by not kicking off the goroutine that waits
for parent cancellation until after the rest of the initialization has
been completed.